server/requirements.txt
flask
flask-cors
google-generativeai
PyGithub
python-dotenv
git+https://github.com/toon-format/toon-python.git